DiveToBall

+++ module/strategy/DiveToBall/src/DiveToBall.cpp
@@ -24,7 +24,7 @@ namespace module::strategy {
         on<Configuration>("DiveToBall.yaml").then([this](const Configuration& config) {
             // Use configuration here from file DiveToBall.yaml
             this->log_level               = config["log_level"].as<NUClear::LogLevel>();
-            cfg.diving_distance_threshold = config["diving_distance_threshold"].as<float>();
+            cfg.diving_distance_threshold = config["diving_distance_threshold"].as<double>();
         });
 
         on<Provide<DiveToBallTask>, Trigger<FilteredBall>>().then(
@@ -37,7 +37,7 @@ namespace module::strategy {
                 // If the distance to the ball is less than the threshold, dive
                 if (ball.rBTt.head(2).x() < cfg.diving_distance_threshold) {
                     // Determine angle to ball and whether we should dive right or left
-                    float yaw_angle         = std::atan2(ball.rBTt.y(), ball.rBTt.x());
+                    double yaw_angle        = std::atan2(ball.rBTt.y(), ball.rBTt.x());
                     BodySide dive_direction = yaw_angle < 0 ? BodySide::RIGHT : BodySide::LEFT;
                     emit<Task>(std::make_unique<Dive>(dive_direction));
                 }
